feat: Add fee history rpc support

+++ src/EtherSharp/Types/TargetBlockNumber.cs
@@ -14,7 +14,7 @@ public readonly struct TargetBlockNumber
     public static TargetBlockNumber Safe { get; } = new("safe");
     public static TargetBlockNumber Finalized { get; } = new("finalized");
     public static TargetBlockNumber Pending { get; } = new("pending");
-    public static TargetBlockNumber Height(BigInteger number) => new($"0x{number.ToString("X").TrimStart('0')}");
+    public static TargetBlockNumber Height(ulong number) => new($"0x{number.ToString("X").TrimStart('0')}");
     public override string ToString() => _value ?? "latest";
 
     public override bool Equals([NotNullWhen(true)] object? obj)
